Set camera translation in Scene

+++ src/com/flume2d/graphics/Graphic.java
@@ -5,7 +5,7 @@ import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 public interface Graphic
 {
 	public boolean isVisible();
-	public void render(SpriteBatch b, int x, int y);
+	public void render(SpriteBatch b, float x, float y);
 	
 	public boolean isActive();
 	public void update();
